Correction du code

+++ tp1/C/analyzer.c
@@ -1,4 +1,4 @@
-#include"analysis/analyzer.h"
+#include"analyzer.h"
 #include<stdio.h>
 #include<stdlib.h>
 #include<math.h>
@@ -7,7 +7,7 @@ analyzer_t * analyzer_create(){
   analyzer_t * res = (analyzer_t *) malloc( sizeof(analyzer_t) );
   res->capacity = 4;
   res->cost = (double *) malloc( sizeof(double) * res->capacity );
-  res->cumulative_cost = (long double *) malloc( sizeof(long double) * res->capacity );
+  res->cumulative_cost = (double *) malloc( sizeof(double) * res->capacity );
   res->cumulative_square = 0.;
   res->size = 0;
   return res;
@@ -26,7 +26,7 @@ void analyzer_append(analyzer_t * a, double x){
     if( a->size >= (a->capacity * 3)/4 ){
       a->capacity *= 2;
       a->cost = (double *) realloc(a->cost, sizeof(double) * a->capacity*2);
-      a->cumulative_cost = (long double *) realloc(a->cumulative_cost, sizeof(long double) * a->capacity*2);
+      a->cumulative_cost = (double *) realloc(a->cumulative_cost, sizeof(double) * a->capacity*2);
     }
     a->cost[a->size] = x;    
     a->cumulative_cost[a->size] = (a->size) ? a->cumulative_cost[a->size-1]+x : x;
@@ -35,25 +35,25 @@ void analyzer_append(analyzer_t * a, double x){
   }
 }
 
-long double get_total_cost(analyzer_t * a){
+double get_total_cost(analyzer_t * a){
   return (a->size) ? a->cumulative_cost[a->size-1] : -1;
 }
 
 
-long double get_amortized_cost(analyzer_t * a, size_t pos){
+double get_amortized_cost(analyzer_t * a, size_t pos){
   if(pos >= 0 && pos < a->size)
     return (pos)? a->cumulative_cost[pos]/pos : a->cumulative_cost[pos];
   return -1;
 }
 
-long double get_average_cost(analyzer_t * a){
+double get_average_cost(analyzer_t * a){
   if(a->size)
     return a->cumulative_cost[a->size - 1]/a->size;
   return -1;
 }
 
-long double get_variance(analyzer_t * a){
-  long double mean, mean_square;
+double get_variance(analyzer_t * a){
+  double mean, mean_square;
   if(a->size){
     mean = get_average_cost(a);
     mean_square = mean * mean;
@@ -62,7 +62,7 @@ long double get_variance(analyzer_t * a){
   return -1;
 }
 
-long double get_standard_deviation(analyzer_t * a){
+double get_standard_deviation(analyzer_t * a){
   if(a->size)
     return sqrt(get_variance(a));
   return -1;
@@ -73,7 +73,7 @@ void save_values(analyzer_t * a, char * path){
   int i;
   if( (f = fopen(path, "w")) != NULL ){
     for (i = 0; i < a->size; ++i){
-      fprintf(f,"%d %lf %Lf\n", i, a->cost[i], get_amortized_cost(a, i));
+      fprintf(f,"%d %lf %lf\n", i, a->cost[i], get_amortized_cost(a, i));
     }
   }else{
     fprintf(stderr, "Could not save values in file %s", path);
@@ -83,6 +83,6 @@ void save_values(analyzer_t * a, char * path){
 void plot_values(analyzer_t * a){
   int i;
   for (i = 0; i < a->size; ++i){
-    printf("%d %lf %Lf\n", i, a->cost[i], get_amortized_cost(a, i));
+    printf("%d %lf %lf\n", i, a->cost[i], get_amortized_cost(a, i));
   }
 }
